+++ framework/base/Security.php
@@ -268,21 +268,21 @@ class Security extends Component
      * [RFC 5869](https://tools.ietf.org/html/rfc5869)
      * @param int $length length of the output key in bytes. If 0, the output key is
      * the length of the hash algorithm output.
-     * @throws InvalidParamException when HMAC generation fails.
+     * @throws InvalidArgumentException when HMAC generation fails.
      * @return string the derived key
      */
     public function hkdf($algo, $inputKey, $salt = null, $info = null, $length = 0)
     {
         $test = @hash_hmac($algo, '', '', true);
         if (!$test) {
-            throw new InvalidParamException('Failed to generate HMAC with hash algorithm: ' . $algo);
+            throw new InvalidArgumentException('Failed to generate HMAC with hash algorithm: ' . $algo);
         }
         $hashLength = StringHelper::byteLength($test);
         if (is_string($length) && preg_match('{^\d{1,16}$}', $length)) {
             $length = (int) $length;
         }
         if (!is_int($length) || $length < 0 || $length > 255 * $hashLength) {
-            throw new InvalidParamException('Invalid length');
+            throw new InvalidArgumentException('Invalid length');
         }
         $blocks = $length !== 0 ? ceil($length / $hashLength) : 1;
 
@@ -316,14 +316,14 @@ class Security extends Component
      * @param int $length length of the output key in bytes. If 0, the output key is
      * the length of the hash algorithm output.
      * @return string the derived key
-     * @throws InvalidParamException when hash generation fails due to invalid params given.
+     * @throws InvalidArgumentException when hash generation fails due to invalid params given.
      */
     public function pbkdf2($algo, $password, $salt, $iterations, $length = 0)
     {
         if (function_exists('hash_pbkdf2')) {
             $outputKey = hash_pbkdf2($algo, $password, $salt, $iterations, $length, true);
             if ($outputKey === false) {
-                throw new InvalidParamException('Invalid parameters to hash_pbkdf2()');
+                throw new InvalidArgumentException('Invalid parameters to hash_pbkdf2()');
             }
             return $outputKey;
         }
@@ -331,19 +331,19 @@ class Security extends Component
         // todo: is there a nice way to reduce the code repetition in hkdf() and pbkdf2()?
         $test = @hash_hmac($algo, '', '', true);
         if (!$test) {
-            throw new InvalidParamException('Failed to generate HMAC with hash algorithm: ' . $algo);
+            throw new InvalidArgumentException('Failed to generate HMAC with hash algorithm: ' . $algo);
         }
         if (is_string($iterations) && preg_match('{^\d{1,16}$}', $iterations)) {
             $iterations = (int) $iterations;
         }
         if (!is_int($iterations) || $iterations < 1) {
-            throw new InvalidParamException('Invalid iterations');
+            throw new InvalidArgumentException('Invalid iterations');
         }
         if (is_string($length) && preg_match('{^\d{1,16}$}', $length)) {
             $length = (int) $length;
         }
         if (!is_int($length) || $length < 0) {
-            throw new InvalidParamException('Invalid length');
+            throw new InvalidArgumentException('Invalid length');
         }
         $hashLength = StringHelper::byteLength($test);
         $blocks = $length !== 0 ? ceil($length / $hashLength) : 1;
@@ -435,17 +435,17 @@ class Security extends Component
      *
      * @param int $length the number of bytes to generate
      * @return string the generated random bytes
-     * @throws InvalidParamException if wrong length is specified
+     * @throws InvalidArgumentException if wrong length is specified
      * @throws Exception on failure.
      */
     public function generateRandomKey($length = 32)
     {
         if (!is_int($length)) {
-            throw new InvalidParamException('First parameter ($length) must be an integer');
+            throw new InvalidArgumentException('First parameter ($length) must be an integer');
         }
 
         if ($length < 1) {
-            throw new InvalidParamException('First parameter ($length) must be greater than 0');
+            throw new InvalidArgumentException('First parameter ($length) must be greater than 0');
         }
 
         // always use random_bytes() if it is available
@@ -551,11 +551,11 @@ class Security extends Component
     public function generateRandomString($length = 32)
     {
         if (!is_int($length)) {
-            throw new InvalidParamException('First parameter ($length) must be an integer');
+            throw new InvalidArgumentException('First parameter ($length) must be an integer');
         }
 
         if ($length < 1) {
-            throw new InvalidParamException('First parameter ($length) must be greater than 0');
+            throw new InvalidArgumentException('First parameter ($length) must be greater than 0');
         }
 
         $bytes = $this->generateRandomKey($length);
@@ -623,20 +623,21 @@ class Security extends Component
      * @param string $password The password to verify.
      * @param string $hash The hash to verify the password against.
      * @return bool whether the password is correct.
-     * @throws InvalidParamException on bad password/hash parameters or if crypt() with Blowfish hash is not available.
+     * @throws InvalidArgumentException on bad password/hash parameters or if crypt() with Blowfish hash is not
+     * available.
      * @see generatePasswordHash()
      */
     public function validatePassword($password, $hash)
     {
         if (!is_string($password) || $password === '') {
-            throw new InvalidParamException('Password must be a string and cannot be empty.');
+            throw new InvalidArgumentException('Password must be a string and cannot be empty.');
         }
 
         if (!preg_match('/^\$2[axy]\$(\d\d)\$[\.\/0-9A-Za-z]{22}/', $hash, $matches)
             || $matches[1] < 4
             || $matches[1] > 30
         ) {
-            throw new InvalidParamException('Hash is invalid.');
+            throw new InvalidArgumentException('Hash is invalid.');
         }
 
         if (function_exists('password_verify')) {
@@ -662,13 +663,13 @@ class Security extends Component
      *
      * @param int $cost the cost parameter
      * @return string the random salt value.
-     * @throws InvalidParamException if the cost parameter is out of the range of 4 to 31.
+     * @throws InvalidArgumentException if the cost parameter is out of the range of 4 to 31.
      */
     protected function generateSalt($cost = 13)
     {
         $cost = (int) $cost;
         if ($cost < 4 || $cost > 31) {
-            throw new InvalidParamException('Cost must be between 4 and 31.');
+            throw new InvalidArgumentException('Cost must be between 4 and 31.');
         }
 
         // Get a 20-byte random string