+++ framework/behaviors/AttributeTypecastBehavior.php
@@ -8,7 +8,7 @@
 namespace yii\behaviors;
 
 use yii\base\Behavior;
-use yii\base\InvalidParamException;
+use yii\base\InvalidArgumentException;
 use yii\base\Model;
 use yii\db\BaseActiveRecord;
 use yii\validators\BooleanValidator;
@@ -217,7 +217,7 @@ class AttributeTypecastBehavior extends Behavior
         } else {
             foreach ($attributeNames as $attribute) {
                 if (!isset($this->attributeTypes[$attribute])) {
-                    throw new InvalidParamException("There is no type mapping for '{$attribute}'.");
+                    throw new InvalidArgumentException("There is no type mapping for '{$attribute}'.");
                 }
                 $attributeTypes[$attribute] = $this->attributeTypes[$attribute];
             }
@@ -255,7 +255,7 @@ class AttributeTypecastBehavior extends Behavior
                 case self::TYPE_STRING:
                     return (string) $value;
                 default:
-                    throw new InvalidParamException("Unsupported type '{$type}'");
+                    throw new InvalidArgumentException("Unsupported type '{$type}'");
             }
         }