+++ framework/helpers/BaseHtml.php
@@ -8,7 +8,7 @@
 namespace yii\helpers;
 
 use Yii;
-use yii\base\InvalidParamException;
+use yii\base\InvalidArgumentException;
 use yii\db\ActiveRecordInterface;
 use yii\validators\StringValidator;
 use yii\web\Request;
@@ -2041,14 +2041,14 @@ class BaseHtml
      * If `$attribute` has neither prefix nor suffix, it will be returned back without change.
      * @param string $attribute the attribute name or expression
      * @return string the attribute name without prefix and suffix.
-     * @throws InvalidParamException if the attribute name contains non-word characters.
+     * @throws InvalidArgumentException if the attribute name contains non-word characters.
      */
     public static function getAttributeName($attribute)
     {
         if (preg_match('/(^|.*\])([\w\.]+)(\[.*|$)/', $attribute, $matches)) {
             return $matches[2];
         } else {
-            throw new InvalidParamException('Attribute name must contain word characters only.');
+            throw new InvalidArgumentException('Attribute name must contain word characters only.');
         }
     }
 
@@ -2064,12 +2064,12 @@ class BaseHtml
      * @param Model $model the model object
      * @param string $attribute the attribute name or expression
      * @return string|array the corresponding attribute value
-     * @throws InvalidParamException if the attribute name contains non-word characters.
+     * @throws InvalidArgumentException if the attribute name contains non-word characters.
      */
     public static function getAttributeValue($model, $attribute)
     {
         if (!preg_match('/(^|.*\])([\w\.]+)(\[.*|$)/', $attribute, $matches)) {
-            throw new InvalidParamException('Attribute name must contain word characters only.');
+            throw new InvalidArgumentException('Attribute name must contain word characters only.');
         }
         $attribute = $matches[2];
         $value = $model->$attribute;
@@ -2113,13 +2113,13 @@ class BaseHtml
      * @param Model $model the model object
      * @param string $attribute the attribute name or expression
      * @return string the generated input name
-     * @throws InvalidParamException if the attribute name contains non-word characters.
+     * @throws InvalidArgumentException if the attribute name contains non-word characters.
      */
     public static function getInputName($model, $attribute)
     {
         $formName = $model->formName();
         if (!preg_match('/(^|.*\])([\w\.]+)(\[.*|$)/', $attribute, $matches)) {
-            throw new InvalidParamException('Attribute name must contain word characters only.');
+            throw new InvalidArgumentException('Attribute name must contain word characters only.');
         }
         $prefix = $matches[1];
         $attribute = $matches[2];
@@ -2129,7 +2129,7 @@ class BaseHtml
         } elseif ($formName !== '') {
             return $formName . $prefix . "[$attribute]" . $suffix;
         } else {
-            throw new InvalidParamException(get_class($model) . '::formName() cannot be empty for tabular inputs.');
+            throw new InvalidArgumentException(get_class($model) . '::formName() cannot be empty for tabular inputs.');
         }
     }
 
@@ -2141,7 +2141,7 @@ class BaseHtml
      * @param Model $model the model object
      * @param string $attribute the attribute name or expression. See [[getAttributeName()]] for explanation of attribute expression.
      * @return string the generated input ID
-     * @throws InvalidParamException if the attribute name contains non-word characters.
+     * @throws InvalidArgumentException if the attribute name contains non-word characters.
      */
     public static function getInputId($model, $attribute)
     {