+++ framework/rbac/PhpManager.php
@@ -8,7 +8,7 @@
 namespace yii\rbac;
 
 use yii\base\InvalidCallException;
-use yii\base\InvalidParamException;
+use yii\base\InvalidArgumentException;
 use Yii;
 use yii\helpers\VarDumper;
 
@@ -165,14 +165,14 @@ class PhpManager extends BaseManager
     public function addChild($parent, $child)
     {
         if (!isset($this->items[$parent->name], $this->items[$child->name])) {
-            throw new InvalidParamException("Either '{$parent->name}' or '{$child->name}' does not exist.");
+            throw new InvalidArgumentException("Either '{$parent->name}' or '{$child->name}' does not exist.");
         }
 
         if ($parent->name === $child->name) {
-            throw new InvalidParamException("Cannot add '{$parent->name} ' as a child of itself.");
+            throw new InvalidArgumentException("Cannot add '{$parent->name} ' as a child of itself.");
         }
         if ($parent instanceof Permission && $child instanceof Role) {
-            throw new InvalidParamException('Cannot add a role as a child of a permission.');
+            throw new InvalidArgumentException('Cannot add a role as a child of a permission.');
         }
 
         if ($this->detectLoop($parent, $child)) {
@@ -254,9 +254,10 @@ class PhpManager extends BaseManager
     public function assign($role, $userId)
     {
         if (!isset($this->items[$role->name])) {
-            throw new InvalidParamException("Unknown role '{$role->name}'.");
+            throw new InvalidArgumentException("Unknown role '{$role->name}'.");
         } elseif (isset($this->assignments[$userId][$role->name])) {
-            throw new InvalidParamException("Authorization item '{$role->name}' has already been assigned to user '$userId'.");
+            throw new InvalidArgumentException("Authorization item '{$role->name}' has already been assigned to "
+                . "user '$userId'.");
         } else {
             $this->assignments[$userId][$role->name] = new Assignment([
                 'userId' => $userId,
@@ -406,7 +407,7 @@ class PhpManager extends BaseManager
         $role = $this->getRole($roleName);
 
         if (is_null($role)) {
-            throw new InvalidParamException("Role \"$roleName\" not found.");
+            throw new InvalidArgumentException("Role \"$roleName\" not found.");
         }
 
         $result = [];
@@ -646,7 +647,8 @@ class PhpManager extends BaseManager
     {
         if ($name !== $item->name) {
             if (isset($this->items[$item->name])) {
-                throw new InvalidParamException("Unable to change the item name. The name '{$item->name}' is already used by another item.");
+                throw new InvalidArgumentException("Unable to change the item name. The name '{$item->name}' is "
+                    . "already used by another item.");
             } else {
                 // Remove old item in case of renaming
                 unset($this->items[$name]);