+++ tests/framework/base/SecurityTest.php
@@ -873,7 +873,7 @@ TEXT;
 
     /**
      * @dataProvider randomKeyInvalidInputs
-     * @expectedException \yii\base\InvalidParamException
+     * @expectedException \yii\base\InvalidArgumentException
      */
     public function testRandomKeyInvalidInput($input)
     {