+++ tests/framework/i18n/DbMessageSourceTest.php
@@ -110,7 +110,7 @@ class DbMessageSourceTest extends I18NTest
     }
 
     /**
-     * @throws \yii\base\InvalidParamException
+     * @throws \yii\base\InvalidArgumentException
      * @throws \yii\db\Exception
      * @throws \yii\base\InvalidConfigException
      * @return \yii\db\Connection