+++ tests/framework/i18n/FormatterDateTest.php
@@ -45,7 +45,7 @@ class FormatterDateTest extends TestCase
         $this->assertSame(date('M j, Y', $value), $this->formatter->format($value, 'date'));
         $this->assertSame(date('M j, Y', $value), $this->formatter->format($value, 'DATE'));
         $this->assertSame(date('Y/m/d', $value), $this->formatter->format($value, ['date', 'php:Y/m/d']));
-        $this->setExpectedException('\yii\base\InvalidParamException');
+        $this->setExpectedException('\yii\base\InvalidArgumentException');
         $this->assertSame(date('Y-m-d', $value), $this->formatter->format($value, 'data'));
     }
 
@@ -502,7 +502,7 @@ class FormatterDateTest extends TestCase
     {
         return [
             ['2015-01-01 00:00:00', '2014-13-01 00:00:00'],
-            [false, 'asdfg', 'yii\base\InvalidParamException'],
+            [false, 'asdfg', 'yii\base\InvalidArgumentException'],
 //            [(string)strtotime('now'), 'now'], // fails randomly
         ];
     }