+++ tests/framework/rbac/PhpManagerTest.php
@@ -129,7 +129,7 @@ class PhpManagerTest extends ManagerTestCase
     }
 
     /**
-     * @expectedException \yii\base\InvalidParamException
+     * @expectedException \yii\base\InvalidArgumentException
      */
     public function testOverwriteName()
     {